Move blob url creation into kolibri-zip package.
