+++ packages/kolibri-zip/src/mimetypes.json
